backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.36.3
sqlmodel==0.0.25
pydantic==2.10.6
python-multipart==0.0.20
better-exceptions==0.3.3
python-dotenv==1.0.1
asyncpg==0.35.0
alembic==1.18.0
psycopg[binary]==3.2.5
pydantic-settings==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0